Add P1_deck_edit and split functions
Cards are now saved in a .txt (csv) file rather than in RAM. So if you change a character in the .txt file it saves next time you open program. There is a button which demonstrates this by chaning 3rd card in deck to Ahsoka.

Instead of Battle1:update() etc in homestate it loops through dictionary called buttons, where buttons are now stored.