Fixed crash when no user data and mouseLastX/Y not being set
In main.lua if P1_deck_file doesn't exist it is created (with every card being equal to '', to stop crashes when it doesn't exist. Now in GameState.lua  and P1_deck_edit cards are only added to the lua table if they're not equal to '' There is a new function for reading player 1s deck, which is called in both of these cases.

mouseLastX and mouseLastY are set to 0 in love.load so if they're not equal to nil, which can cause crashes.

Co-Authored-By: James <36278229+FormularSumo@users.noreply.github.com>