Add pause button and split update/move/attack into 3 different functions
Now move is called for each card before attack is.

Buttons are now rendered only if their variable called should_render is the same state as the state function calling it. So homestate and gamestate both call the update and render function for all buttons, but those commands are only carried out if the they're supposed to.

Pause button allows the user to pause a battle mid-way through. In future will probably be the beheaviour if you bring up a menu, but for it's useful otherwise and useful for me trying to bug test things.