Re-name self.attack to self.offense
This is primarily because each card has a function called attack. But previously the function had to be called something else (attack1) because it had the same name. Offense is only a noun, and the opposite to defense, while attack is also an action, so offense seems like the better word. Now the attack function of a card is called Card:attack