Improve deck saving
Previously an extra comma was added each time Player 1 deck.txt was written to, a very small memory leak I suppose. It had increased my file size from <1 KB to 4.

Also instead of checking if Player 1 deck.txt file doesn't exist before writing an empty deck to it (for reading purposes) it now checks if the file is empty. This means a blank deck will be written if the file is empty as well as if the file doesn't exist