Fix Imperial March music glitch on Android and fix mute key
Imperial march piano only.oga was playing once through like it should. However Imperial march duet was playing a new version from the start every time the program was opened from multi-tasking view. This was because the 1st imperial march was being paused automatically by love2d and the 2nd one was playing when the game was loaded again because the 1st had stopped playing - the condition for the 2nd one to play. The 2nd one only plays now if the window is in focus - meaning it won't start playing until the 1st one has actually stopped.

One of the earliest things I added to the game was that the m key would pause/play music in a battle. That stopped working when I added the pause button because the pause button played music when unpaused, and paused music when paused, over-ridign the m key. Now the m key sets volume to 0/1, so functions again, and across the whole program in the way a mute key would be expected to function. I will probably add a button for it at some point (for touchscreen, and so people know it's an options), probably whenever I get round to making a settings menu...