Winner is printed after game and :update() doesn't run
also re-named render_gamestate yet again, this time to renderstate. Because gamestate is the name of a specirfic state.

When P1_deck or P2_deck is empty, the table is deleted and the winner variable set accordingly. The winner is then printed and the gamestate:update function stops running. Background video and audio keep playing. Hitting pause will pause/unpause them.

Removed some old commented test code that printed cards alive in a players deck, which has now been adapted to work out if a deck is empty and therefore needs deleting.