Add gamespeed slider and cahnges to Button.lua
Gamespeed slider slides between 0.01 (minimum value allowed) and 1 (maximum value allowed. This number is then multipled by 4 to get gamespeed because I wanted to gamespeed to be able to speed up as well as slow down. There's a slightly 'lock' on 1x gamespeed, where a few percentages to either side of the value automatically become 1x. This helps with figuring out where 1x game speed is on the slider.

The button table has been re-named gui and now contains sliders (1, so far). Now gui elements are created in the state:init functions and the table is wiped clear upon exiting a state. This means all gui updating/rendering can be done in main.lua, and there doesn't have to be a check to see if the gui element should be rendered in this state. So the renderstate variable has now been deleted. Doing things this way means only the buttons that are in use are stored in RAM and less code is needed. It does probably mean ever so slightly longer load times when switching between gamestates.

The slider is a class which functions quite like my button class, meaning it's very easily to make new sliders in future. Moved pause and prebuilt deck to a file called other_functions. The gamespeed_slider function is also in this file.