Stop escape key exiting fullscreen - broken on Android
https://love2d.org/forums/viewtopic.php?t=82762

Back key on Android is binded to escape, meaning that when you clicked escape on Android it exited fullscreen. Problem was you can't click F11 on mobile (if you do plug a keyboard in it crashes for some reason...) to go back into fullscreen, and more over you're not really suppoesd to exit fullscreen on mobile as there's not really any point being in windowed mode. Now clicking back key/escape doesn't do anything... in future it will bring up a menu though.