Remove love.touchpressed function, move mute code & add comments
Removed love.touchpressed function as it didn't do anything. I added it because I thought it was needed to record when someone was touching the touchscreen, for animations. But actually if it was being used for that the buttons would grow for a split second then shrink again while holding down on them. The reason the buttons were staying big while touchdown was because love.mouse.isDown(1,2,3) is triggered by touchscreens as well as mouses. So now only that is being used, rather than that in addition to love.touchpressed.

Functions exactly the same as before but performance should be a tiny bit better and less code is used now. Love.touchreleased is still used for actual clicks.

Code that checks if 'm' is pressed (for mute) was previous in love.update and used global table. I moved the code from gamestate to main a while ago which is why it's in love.update. But if it's always being checked for it's more efficient to have it in love.keypressed function so it's only checked when a key is pressed rather than being checked every frame.