Add basic controller suppport - left joystick/d-pad moves mouse, A left clicks
There's no right click bind but there's no right click in the game - as the game is supposed to be fully/easily playable on touchscreen, so that's not an issue.

This took a while and was inspired by me being asked to test a controller with the LOVE2D 11.4-preview patch that fixed USB devices crashing LOVE2D on Android (see https://github.com/FormularSumo/Star-Wars-Force-Collection-Remake/issues/22).

Controller stuff is handled by LOVE2D so should work with any controller/joystick and any platform. I've only tested on an Xbox One controller on Windows/Android though