Improve mouse position updating code
Instead of having the same mouse position (mouseLastX, mouseDown etc) code multiple times I've made a function in other functions that does that and is called.

I've also removed the mouse updating code from all of the released function, as there's no need to update the mouse position when it's already being updated in the love.mouse.isDown code. in love.update().