Add pause_on_lose_focus setting, song queue manager & improve button.lua
When true (which it is by default) game is paused when it loses focus, and unpaused when it gains focus. This was mostly done because on phones you don't expect games to carry on playing (and music to carry on) when you exit the app. On desktop/laptop you might, and on either platform there's a button to toggle it.

Also there's a universal music manager in the main love.update() function. So when loading a state songs0], songs[1] and so on are defined, the number being the order they should be played. A variable called queue length is also defined and then the song queue manager does the rest. Songs resume where you left of if you pause/unpause, and when the queue is empty no more songs play. It also doesn't crash if you don't define any songs or queue_length

Songs table was previously called sounds, but now it's called songs because it's the song queue. Maybe I'll make another table called sounds... when I put in some game sounds (eg blaster firing, lightsaber swinging etc)!

Buttons now have a function called Button:update_text() which when called with a string argument, updates the buttons text to be that. You could just manually change the button's text, but doing it this way means the button will automatically re-allign by re-centering if applicable and re-centering with image if applicable.

I also reduced grow on button click from 10% to 8%.