Add Endor background and improve video player
https://pericror.com/zoom-background/star-wars-zoom-background/
Endor background isn't used yet but works fine if I manually set it. Looks good, except button text will need to be white.
Videos have been moved to Backgrounds folder and videos folder deleted. All backgrounds are saved as the same name now when they're created, and with the help of a couple other variable in the same table (background type - video/photo and seek - for where to rewind back) a single background manager in main.lua manages all backgrounds. It pauses/plays videos on loose focus (if setting is on), loops when they reach end of video and draws them.