P2_deck_cards is cleared upon GameState:exit so previous levels don't affect new ones.
Eg if you played Battle1, went to menu, then clicked on Throne Room battle all cards not set by throne_room (all but middle front 4) would be same cards as battle1, which is not what was supposed to happen. Clearing P2_deck_cards in GameState seems to easiest way to do this - it'll automatically be cleared for all new levels and frees up (a miniscule amount of) RAM when you leave GameState.