Fix bug causing ranged characters to sometimes not fire
I deleted an optimisation that was supposed to check if closest enemy character to you (ie one in same row first column of enemy deck) was out of range and if so stop card checking if any other enemy is in range however it was actually if opposite character was out of range don't bother checking which meant if Rex was number 15 back row he would check if number 15 enemy back row was out of range and if so (which is so) didn't fire. I could probably fix the optimisation but that would reduce how effective it is and performance is so good anyway I don't think it really matters.