Fix cards moving forward twice per turn and not row shifting when row isn't full
First problem fixed by using an elseif. So now card can either move forward because it's not in forward position OR because the card that should be in front is dead, not both.

Second problem fixed by moving x = x + 6 out of if card exists loop. Previously row would only shift upwards/downwards if the first card in the row existed, which was ok but not ideal. Now rows will shift whichever characters exist upwards/downwards by checking each character regardless of if first in row exists.