Hide pause_on_loose_focus settings if OS = Android and increase font size/positioning
On Android (and iOS) it doesn't matter if the game tries to continue in background or not; it won't, or at least not properly. Music is paused and battles (movement) can mess up.

Increased font size on pause on loose focus and create prebuilt deck button to 60 and moved them down a little - pause on loose focus is in line with volume slider now (which is now centred between that button and left of screen)