Check if mouse has moved from where it snapped to rather than waiting on timer to make mouse visible
Old method didn't work if device took longer than 0.02 seconds to load new scene (as my laptop does for loading games). That timer could be reduced but then there's a visible delay between moving mouse and it becoming visible if mouse has just snapped to a button. Instead of having to have a trade of between those 2 things this method works 100% of the time on any device without any delay.