Call love.timer.step at end of state transition
This resets the timer and therefore dt. My reasoning for doing this is that previously while a new state was loading (pictures, video etc) love.run wasn't running so dt kept getting bigger and bigger. The problem with this is that my controller (and probably most) have a very slightly lean in one direction. So when love.run was finally called again all that dtr multiplied by controller lean registered as the gamepadaxis having moved significantly, enough to bypass the deadzone and move the mouse, hence making it show, exiting mouse snapping mode and even making the mouse not snap to the right button sometimes.