Improve smoothness and performance of sliders
Now when both arrow keys are held down self.held_time is set to 0 so there's no janky teleporting of circle when you switch direction. And mouseposition is updated as slider moved, so if you move mouse off it while holding down arrow key (say you held down to end and didn't take finger off) mouse still comes out of circle.

Additionally there's less if statements being checked at any one time and self.held_time is not constantly being set to 0.5. And Slider left/right key down code is now in slider.lua rather than state files.

Reposition mouse now checks if index is itself a GUI element (if it's not an index to one), and sets mouseTouching to whichever of those is true (ends function if neither).  This means you can now feed reposition_mouse a GUI element as well as an index to one, which was needed in slider.lua (as sliders don't have their number as a variable). I could have made it a variable and passed it to it, or worked out in init but