When checking if mousemoved allow up to 0.5 pixels movement before triggering
On Android 11 decimal point float is returned and it's not accurate enough to use that as it sometimes doesn't align with previous value, usually by up to ~0.25. So there has to be a bit of a deadzone, which I've set to 0.5 as that's how far Windows can be off because it automatically flots x and y in mousemoved.

I've also made it so either x or y can have moved to trigger movement, previously it had to be both but it's accurate enough now to have separate, which makes moving mouse to make it show feel much more fluid.