Fix crash due to double row movement and some movements being delayed
Now a table which contains movement to be carried out for each row is updated, and then movements are carried out based on results. Ensures only one movement per row is ever executed. Means the return statements within the main movement code can be moved which fixes row 0 not moving when 3/4/5 were.