Improve characterStrength function/formula
If range is 1, then meleeOffense * 0.9 is used as Offense

If range is > 1, if meleeOffense is higher than ranged then meeleOffense * 0.9 + rangedOffense * 0.1 * a very close to 1 range modifier is used. The range modifier is to ensure that higher ranges still result in higher stats. However for a more melee focused character like this, having more range likely isn't going to be useful - unless the character isn't put in the first row, only at the start of battles at approach and later if there's no adjacent characters to attack.

If range > 1 but ranged offense is higher, than the contribution of ranged offense (added to that previously mentioned multiple * 0.1 for consistency, so that having a higher ranged offense can't result in lower overall power) is dependent on range, ranging from ~37% (range 2) to ~64% (range 5). So characters more ranged attack focus do benefit quite a lot from having a greater range. However, for all characters melee is still important, as if a battle is close it will always come down to melee for all characters for one side at least.

The idea behind doing this was to more accurately reflect how cards would be used in formations, and better reward cards which have a higher range where appropriate.