Fix evade not being used correctly in character strenth
I assume the 1/2 should have been in brackets or written as a 0.5. As it was, it was just being multiplied by the 2 and doing nothing. What I imagine I intended and have done now is made it exponential, so more evade increases overall strength quicker the more it goes up, as dodging an attack means your defence is kept higher for longer benefitting disproportionately, and also low evades are relatively likely to not ever be triggered in the course of a battle.