Fix arrow keys adjusting gamespeed and volume in-game
The old code in slider set default to true if it was nil or false, meaning it was always. true.

Additionally in gamestate default was not being changed for the volume slider when un/pausing. I moved all the default code below the text updating for consistency.

Also the gamespeed slider was being set to visible when pausing which wasn't needed as it should always be visible