Cards images used in battles are loaded off the main thread
Also change imagteDecoderWorking channel to have 1 pushed to it instead of "working", as the actual value is irrelevant there just needs to be one. 1 (or TRUE) is the pretty much smallest value it can be, and made variables in imageDecoderThread local.

Unfortunately cardImages can no longer be deleted when the battles has ended as they're no longer stored by the Cards themselves, so would cause a crash. It would be possible to save them to cards when they're retrieved by the main thread but probably more effort than it's worth. Overall memory used in battles has increased by about 1.5 MB (both in battle and after finished) which is unfortunate but not loads.