Rewrite code for decoding images off thread, and extend to projectiles and weapons (in battle)
Previously images would get loaded multiple times (though not stored) if there were multiple copies of a card (or projectile/weapon) loaded at the same time. Now there's a gamestate table for recording all the images that need loading, then loads one copy of each and updates the image variable for the relevant cards/projectiles/weapons. This means we're back to using local variables for images as well as a gamestate table (as opposed to global table), so after a battle has ended the images no longer in use can be deleted straight away. It should also be faster because local variables. Additionally, image names no longer get stored for cards/projectiles/weapons, just the actual pointer to their image (once it's been loaded). Memory usage in battles is now almost dentical to pre-multithreading work (technically a tiny bit better, somehow), whereas before it was a little higher.

I've also decided not to show any placeholder card if a card hasn't loaded, but rather nothing. If this ever were to actually be displayed (incredible unlikely), it should still be clear that it's lodaing because normally there would be a blank card in any empty spaces. This means that the blank card can be loaded off-thread as well now which technically improves deck editor load time by like 2%

Another advantage to this is that some of the initalisation logic for projectiles and weapons it deferred until the image is ready, and so should be done at less busy point in time (not the initial battle load run).

In order to avoid sending extra info over channels, I put all images types (card/projectiles/weapons) in the same table. The only disadvantage to this is that you couldn't have any of these things have the same name, but I don't think that should ever be an issue.

Also deleted some nil assignments for local variables that were unnecessary.