Move away from using imageDecoderWorking channel.
Now that pushes to the imageDecoderQueue are done together after creating objects (to ensure duplicate images aren't loaded twice), there's not really any advantages to starting the threads directly before this. Not using imageDecoderWorking allows the code for starting and managing thread decoding to be simplified, which reduces the main thread overhead of pushing and popping to this channel (about 5 ms when swapping cards for example). It should also reduce how much the imageDecoderThreads need to access this channel (as they no longer peek first, just pop, and also don't spend time waiting for things to be pushed to the queue), although I haven't tested the performance impact of this.