Keep P1deckCards always stored.
As it's used in gamestate and deckeditorstate, and loaded in many cases in load.love anyway, it makes sense to just keep it stored. Especially as it takes up a tiny amount of data just being a key value table, but a non-negligible amount of time to load given it's file reading. Now it's only loaded during startup and otherwise just saved when necessary. I've removed it from being a condition for tutorial to run given that it doesn't automatically get saved now, if for example you were to delete the file manually.