Replace Starry Sky picture background with higher quality one.
I directly exported a png from the source video in Shotcut which looks original quality. Subsequently converting it to a jpg made it bigger and look worse, and exporting as a jpg in Shotcut caused it to miscolour (and still look worse otherwise/be biggeR), so I've left it as a png. It's actually smaller as a png than the old jpg and any of the other backgrounds so I'd say this is a good result.