Fix offscreen cards being created too late
Offscreen cards are supposed to be created 1 second before of going onscreen, which was happening for the first 7 seconds of the game, but then after that they were being created immediately before going on. I only noticed because checking for decoded each movement was not loading the decoded images in time. Now images have a second time-window to be decoded in, which should be plenty given they seemingly had almost no time previously and I didn't notice.