Switch to using love.filesystem to write/read data so it persists across reloads.
For some reason binser when using io.open() and file:write() save files in the same directory as the game code, which means it doesn't save when you leave the website. Not sure what would happen if I fused the game and ran it natively. Using love.filesystem it saves in the normal love save directory and so uses indexedDB on love.js to persist on the user's browser.