All states use enter instead of init
For consistency. Also in theory init is only run when a state is initialised, whereas enter is every time they're run. In practice, my statemachine library always deletes the old state when it changes state so it doesn't really matter but good to be correct. In the past it's also caused issues where init didn't have parameters but I wanted certain code to be run earlier (eg offthread stuff), so good just to have it all in a function with the option of parameters. Also moved gui creating to end of homestate like other states