Fix crash toggling background video
Due to Background['Name'] not being saved as full path now, confusing the updateBackground function that checks if a file with that name + folder + .ogv exists. Now Background['Name'] has been flipped back to being just the name, and Background['Filepath'] exists for the full path. updateBackground now just gets background info from backgroundInfo function now which is better as less duplicated code and eg if I change video format in future if Love2D 12 supports it, only need to update in one place.

Also changed seek to be returned as nil if not video as not used, and HomeState not to save it as not used there.