Rewrite draw code, use canvas for card viewer background
There's no longer any state-specific code in love.draw or other random places, instead it is now all contained in the states themselves. This has the additional advantage of not having canvases around when not needed, and using a local one in deckeditstate (gamestate this breaks moonshine for some unknown reason). Have renamed renderBackground to renderNormal and added a new renderBackground which is actually for rendering the background should a state want to.

Card viewer background canvas has reduced draw calls from ~25 to 11.

Also moved collectgarbage() to end of StateMachine:change, when the old state should no longer have a reference to it. Unfortunately still doesn't seem to be clearing all the old state's data.