Fix crash if swapping deck card with blank in inventory
Regression in 0.12

The code was incorrectly checking if the card in the deck was blank when swapping with a blank card in inventory, rather than other way round. I've also moved the update/delete evolutions into this check as although they don't crash due to checking if not blank in the functions, no need to do this if we already know they're blank so won't have evolutions.