Animation when card can be moved into blank slot
Previously blank cards would never have the highlighted animation because they can't be selected/dragged around. Now they have the animation only if another character is being moved and so they can be swapped. This doesn't apply to blank cards in inventory as they generally shouldn't be swapped with (inventory cards don't get swapped, remove button is for removing deck cards). This also doesn't affect the keyboard/controller behavior which was to already do this.