Optimise Character png size using Imagine
Characters folder: 52.8 MiB --> 42.9 MiB (-10.1 MiB) = -19%. Note: this is *including* the larger jpg images, which were on average around twice as big as the smaller pngs. This means the actual reduction in size to the pngs is around 60-70%!

If I compare some cards in an external viewer blowing them up to fill me screen, find a difference and then look *extremely* closely in-game while constantly flicking back and forth between game versions (pre/post this commit), I can _just_ about see a tiny difference. But I think that's easy worth it for the file size reduction.